Group location field
add a group location field(set), that defines the geographical location
of a user group. Data requirements:

- locality
- locality type (city, region)
- country
- longitude, latitude coordinates

long, lat required for displaying user-group on local map
the global map view will be use the country's central coordinate.
the country field must be linked to another dataset where all countries must
be linked to a continent. (it is required for group by continent view)

Change-Id: I373ed07615cd55054770073c63b5034cd56796c7