Add release-manifest tool to scripts
This tool helps to build a standard module manifest file that
holds information about production and development releases of
the groups portal distribution, separated from drupal.org module
repository. The jenkins release jobs can call this script as the
part of the release process and generate the recet manifest information.
The staging and production instances can fetch updates based on
this xml catalog.

Change-Id: I9b9be5ad4c74dca192dc966f450fa45fa54566e9