New bootstrap based theme
Add a bootstrap css based theme to groups portal, and match the design
elements and layout of new openstack.org site. This patch will not
enable the theme by default, just add the bootstrap module and the
openstack subtheme. The subtheme is based on the sass version of
bootstrap with a bundle compass support.

Change-Id: Iae976151e87d4d2edd6e913a5aea16d9c13dc051