Follow profile picture changes in openstack.org profile
Until the openstackid.org oauth2 not returns the hash of the profile picture
or a ttl of the last change that auth module can check, the pictures
expiring in every 24h. TTL time is stored in user profile as oauth2_picture_ttl
parameter, and authentication retriggers user profile save hooks when the
ttl is expired, resulting of refetch of the user's picture.

Change-Id: I01c84864d61248bf45932d21c21f5b62983d5df2
StoryId: 2000114