Add other-requirements.txt
Add other-requirements.txt to install only required
packages that we need for testing and not extra packages
that are installed by default.

This allows to remove the php packages from the global bindep-fallback
list that is installed on every test node.

Node that this does not change what is installed for running groups,
this is done in the manifest.

Change-Id: Ic175744997ac01b3c89793c78d501d43044507e8