Group contanct report csv dump filter by group status
The reports/group-contact-report/csv export now filterable by the
group status using the ?status parameter. The format is:

https://<baseurl>/reports/group-contact-report/csv?status=official

Valid values:
- official: filter the official groups only
- unofficial: list the non-official groups

When the status parameter is not present in the query, the
resultset will contain all the user group independenlty of
the group's status.

The reports/group-status-report also includes this url filter,
and the title contains the official status.

Change-Id: I8fcfe6e9963c8b4bbfaee6e1231acc1227ff37c7